[Silver II] Title: A → B, Time: 232 ms, Memory: 19180 KB -BaekjoonHub

+++ 백준/Silver/16953. A → B/A → B.java
@@ -21,7 +21,6 @@ class Main {
         for (int i = 0; i < 10; i++) {
             set[i] = new BitSet();
         }
-
         Queue<Node> q = new ArrayDeque<>();
         q.offer(new Node(N, 1));
         long result = -1;
@@ -31,10 +30,9 @@ class Main {
                 result = n.cost;
                 break;
             }
-            long next = n.no * 2;
-            int hash = (int) next / 100_000_000;
-            int key = (int )next - (hash * 100_000_000);
-//            System.out.println(hash + " : " + key);
+            int next = n.no * 2;
+            int hash =  next / 100_000_000;
+            int key = next - (hash * 100_000_000);
             if(next <= M && !set[hash].get(key)){
                 set[hash].get(key);
                 q.offer(new Node((int)next, n.cost + 1));
@@ -42,14 +40,11 @@ class Main {
             if(n.no > 100_000_000) continue;
             next = n.no * 10 + 1;
             if(next > M) continue;
-//            System.out.println(n.no);
-//            System.out.println(next);
-            hash =(int) (next / 100_000_000);
-            key =(int) (next - (hash * 100_000_000));
-//            System.out.println(hash + " : " + key);
+            hash =  next / 100_000_000;
+            key = next - (hash * 100_000_000);
             if(!set[hash].get(key)){
                 set[hash].get(key);
-                q.offer(new Node((int)next, n.cost + 1));
+                q.offer(new Node(next, n.cost + 1));
             }
         }
         System.out.println(result);
